Confirm update

+++ Bin/00updatedata.sh
@@ -1,4 +1,9 @@
 #!/bin/sh
 
-rsync -av --delete /mnt/hdd02/abx/data ./
+read -p "Are you sure? " -n 1 -r
+echo    # (optional) move to a new line
+if [[ $REPLY =~ ^[Yy]$ ]]
+then
+	rsync -av --delete /mnt/hdd02/abx/data ./
+fi
